Flyway integration

+++ src/main/java/com/fict/repository/TestRepository.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
  * Created by stevo on 4/27/17.
  */
 @Repository
-public interface TestRepository extends CrudRepository<TestEntity, Long> {
+public interface TestRepository extends CrudRepository<TestEntity, Integer> {
 
-    TestEntity findTestEntityById(Long id);
+    TestEntity findTestEntityById(int id);
 
 }
